Fix: use temporary workaround for test compilation failures
For some reason, using the `futures-channel` crate over the `futures`
crate causes compilation errors. This is very strange because the
`futures::channel` module from the `futures` crate is literally just
re-exported from the `futures-channel` crate!

Despite this, the code fails to compile whenever I use `futures-channel`
instead. As a temporary workaround, we use the full `futures` crate in
the tests. However, when the user downloads the crate, they will still
get the minimal sub-crates (not the full `futures` crate).